Backport BPGES settings markup from plugin into citytech theme
The code used to create the Default Settings markup in BPGES is sensitive to
a null value standing for "No Email". That markup had been copied to the
citytech theme before this change took place in GES, with the result that
the citytech theme markup (Group Settings > Settings) was not properly
displaying the No Email setting.

See #826