Switch openlab_delete_group redirector to groups_group_deleted
This function was previously hooked to groups_before_delete_group, which meant
that the user was redirected before the group was actually deleted, and before
any of the functions hooked to groups_before_delete_group had a chance to run
their logic. This includes openlab_delete_portfolio(), which wasn't getting a
chance to remove the user portfolio metadata.

See #878