Skirt bp_new_group cookies when setting up group creation page
The bp_new_group cookies are not being properly deleted in all cases. As
a result, when you try to create several groups in a row, incorrect info
shows up in some of the fields. This is most problematic with the course
site info, which can be pulled from the wrong group so that you see the
"unlink" interface incorrectly.

The current changeset does two things:
1) Manually ensures that the "unlink" interface can never show up for
the group create page
2) Attempts to clear the new_group cookies when looking at the
group-details page. This part of the change doesn't catch every case,
because it doesn't run early enough (and can't), but it will catch some
edge cases.

See #846