Fixes comment privacy in Grader plugin
The Grader plugin was using the 'comments_array' filter to protect
@grade comments from appearing to non-authorized users. However, this
technique only works when 'comments_array' is called, which only happens
when the comments template is called. Thus, direct calls to comments -
such as those using get_comments(), like on Recent Site Comments widgets
- miss the filter entirely.

We work around the problem by filtering at the_comments as well.

Note: This is only a reliable method if Grader is also activated on
the main site, where it'll be able to catch BuddyPress content.

See #949