During course clone, ensure that source group admins are pulled up correctly
A change in BP 2.0 turned 'populate_extras' off by default. As a result,
group admins were not being returned with the group object; in turn, no items
were being matched as "belonging to an admin" during the course clone process,
and thus were not being copied over to the clone destination.

See #1112