Styling and markup updates for Email Members page
This commit involves a 'quick and dirty' solution for injection Bootstrap
markup into the Email Members form. Basically, on the admin.php page
powering the group profile admin pages (a custom theme file), there is a
conditional that uses a custom function for the Email Members page,
ommitting the groups_custom_edit_steps action from this page altogether.
Obviously this is not an ideal solution and will be explore more fully in
a future release, along with other plugin/markup situations.

The custom functions for Email Members are located in the newly created
email-funcs.php file (in the functional library), and are pulled from the
Buddypress Group Email Subscription core