Styling and markup updates for Your Email Options page under Course Profile
See full explanation for what is happening here in commit
3e3e6525187b51c23d59adde95c0e43230afda71

In this particular case, a catch on the homepage is routing the Your Email
Options page to a custom notifications.php file under groups->single. This
file calls a custom function based on a function from the Buddypress Email
Group Activity plugin. The main goal of this custom function is to inject
Bootstrap markup into the Your Email Options form