Style and markup updates for Group profile files page
The markup output for the group profile files page is now generated via
custom functions derived from the Buddypress Group Documents plugin, but
stored in the theme under the files-funcs.php functional library file.
These custom functions exist so that Bootstrap markup and classes can be
injected into the markup for the file page.

It is the hope that future releases of the plugin will provide the
necessary mechanism for hooking into the markup of the file page, so that
we don't need to use the custom functions (which make updating more
difficult).