Fix bbPress new topic posting for non-admins.
There were two problems. One is that non-admins did not get the
publish_topics cap mapped to them, while admins had it by virtue of
is_super_admin(). This is because bbPress maps everything to 'participant',
which we don't want to be tied to. I've remapped to 'exist'. Two, bbPress
uses get_permalink() to populate the 'action' attribute on forms. In BP groups,
get_permalink() returns the URL of the WP page for the groups directory.
bbPress tries to filter this and correct it, but it failed for some reason in
this case. I've corrected it.

See #1191.