Force 'hidden' forums to be 'public'.
I'm not sure what use 'hidden' forums are in bbPress when used inside of BP
groups, but the hidden status causes loads of visibility problems. These hacks
ensure that we query for hidden groups, but then reset them to public before WP
has a chance to intercept. It's really a nightmarish set of hacks.

See #1299.