Update for merge:  Do template_notices hook on change-avatar.php screen.
We need to add this manually because change-avatar.php is not loaded in the
normal BuddyPress member profile environment (we've moved it so that it appears
under the minified header). We need the template_notices for avatar errors.

See #1030.

Original: https://github.com/livinglab/openlab/commit/4839f0335730c7e110626015c5ee71c1801d16ea