Show registration errors when Major is not properly filled in.
The Major Field of Study field is not present when the page is built; instead,
it is populated dynamically via AJAX when the Account Type is selected. This
means that the BP native method of showing error messages doesn't work, since
it relies on WP's hook system, which only works across a single pageload.

As a workaround, we load registration errors into a JS global when the
registration page is built, and use JS to populate the error messages when
fields are generated by AJAX.

See #1385.