Bug #1384 - Mobile: Breadcrumb links missing on footer
Fixed by putting footer markup into a separate function with an available
placholder variable. The placholder variable outputs a set of footer
markup (minus any js) for purposes of vertical spacing; some of the themes
available to group sites have a max width limit set on the body,so the
visible footer is absolutely positioned in order to maintain the
end-to-end appearance across all themes. However, on smaller screen sizes,
the footer naturally grows in height, and yet the main page container has
no idea this is happening. As a result, the footer begins to creep up and
hide page elements. The placholder footer is set to visibility: hidden, so
it will not be visible to the user, yet will provide the page container
the necessary height feedback to keep all page elements above the footer.

This solution is designed to avoid a messier js-based approach. Given the
limited amount of markup in the footer, I don't believe it presents any
substantial addition to page load