Create explicit 'disable' flag for group forums.
The existing `enable_forum` flag is utilized inconsistently in bbPress 2.x and
in the BP implementation of bbPress 1.x, with the result that certain groups
that previously had forum_id and the forum tab enabled no longer have it set
because enable_forum=0. Now it is only possible to disable group forums
explicitly.

See #1387.