Use bp-activity query to get Who's Online user IDs.
Previously, we were querying last_activity in usermeta. This method
of storing user last_activity data is deprecated. It is also very slow,
and difficult to cache. Using the more up-to-date data in the activity
table means that the raw query is faster, and it also means that we
will automatically benefit from the caching layer introduced in
https://buddypress.trac.wordpress.org/ticket/7237.

See #1943.