Fix incorrect 'name' attribute in group category form.
The group taxonomy library in bp-customizable-group-categories detects
the deletion of all categories from a group by comparing the empty
array of categories-to-add against the array of existing-categories.
A mismatch in the 'name' attribute between the view and the controller
meant that this comparison was always failing.

For the record, the '_existing_group_categories' mechanism is probably
overkill. If an empty array of categories is sent, this seems like
enough information to perform a deletion.

See #2083.