Bug #2175 - Accessibility : update labels on 2 contact forms
This issue was primarily resolved by moving the form markup *out* of the admin
and into the codebase, and then directly updating that source code to meet
accesibility standards.

The intro text is still in the admin, so that can be changed, however any
future changes to the form elements will need to flow through development.