Bug #2183 - Accessibility : 2011 Theme repeats search form labels
The main issue here is that the 2011 theme calls get_search_form()
multiple times in one page (inclulding via widgets). The get_search_form()
produces markup using the same "id" attributes everytime, which is semantically
incorrect and fails WAVE accessibility testing.

To fix this:

Server-side

1. First we hook into the get_search_form filter
2. In the hook callback, we check against relevant themes (currently only
set to "twentyeleven", but can be expanded if necessary)
3. Then we check for a global, twentyeleven_search_form_count - if that
global does not exist, it is declared with a value of 1; if it does exist,
the current value is iterated
4. We then use DOMDocument to parse the form HTML
5. During this parsing, all "id" and "for" attributes are updated to include the current value
of twentyeleven_search_form_count
6. The legacy id value for each parsed node is set as a class value
7. The parsed object is saved to the filter value $form, and returned

In this way, every form on the page has a unique set of "id" attributes.

Client-side:

1. Because a number of the styling rules are set to the id attribute of
the search form, we needed to append the search form stlying to account
for the unique ids
2. First we created a stylesheet under the "theme-fixes" folder called
"twentyeleven.css"
3. Then we set this stylesheet to append to the twentyeleven theme styles
via openlab_load_theme_fixes()
4. Next we copy all of the search-related styles from the twentyeleven
stylesheet into twentyeleven.css
5. Finally we update each of those styles to use the classes set in
Server-side#6