Bug #2193 - Accessibility : add alt text to homepage slider to satisfy WAVE checker
The issue here is that the client-side plugin responsible for creating the
home page sliders, Camera JS, does not include an 'alt' attribute on the
images the plugin generates. Camera JS also has limited hooks, but a first
pass at solving this issue attempted to use one of those hooks to inject
the alt tags after the images were loaded to the document. Unfortunately
the timing of this doesn to seem to satisfy the WAVE tools.

Given that Camera JS has not received updates in a number of years, I have
made the decision to fork it inside the theme, indicated by the new file
name "camera.mod.js". Withing this fork, I updated the plugin to include
"alt" attributes, based on the presence of a "data-alt" property in the
source markup.