Bug #2159 - Help section: Issue with child category order
This appears to be a case of some legacy functionality that was never
fully cleaned up when we switched over to the new design. It would seem
the legacy layout for this page involved two columns, and the legacy
functionality was designed to maintain the taxonomy ordering across the
columns. However, when we switched over to the new design, this legacy
functionality was essentially "flattened", withouth removing it
alltogether, and this threw off the taxonomy ordering comeing out of the
tax query.