OpenLab Gradebook Best Practices: Removing BP/OpenLab depdendent fields
This commit removes any references to xprofile fields in OpenLab
Gradebook. A filter has also been added so that OpenLab can re-add the
xprofile fields via mu-plugins.

The xprofile field references have been replaced by a function that
attempts to pull in the user meta fields "first_name" and "last_name"
instead. If those fields are not available, it falls back to parsing the
"nickname" field.