Bug #2209 - Gradebook Cannont Install Plugin & Bug #2210 - Gradebook: Not able to add Gradebooks
These two bugs were related. Fixes include:

1) The $wpdb declaration was missing in some functions
2) Some of the $wpdb->prepare() statements in the Database configuration
were missing the variable argument
3) The plugin did not set up an initial instructor user, so the
permissions check was failing on the setup of an initial Gradebook due to
the permissions check. This was resolved by setting up an initial
instructor user with a gbid (Gradebook ID) of 0. When the first Gradebook
is set up, this initial user's gbid is updated to that first Gradebook's
gbid