Bug #2212 - Gradebook: Not able to add Assignment
This issue occurred whenever an assignment was submtited without all of
the fields filled out - the database table requires non-NULL values. For
now, fallback values have been instituted. A future recommendation is to
add client-side validation so instructors will be aware that fallback
values will be used in absence of manually filled in data.