Bug #2204 - Gradebook: plugin creates menu item on group site
This bug was returned due the original fix not working for the Hero theme.
It turns out the Hero theme uses a custom callback for the menu fallback,
and inside of this fallback, wp_list_pages is utilized to output a
fallback menu. So the custom storage page was added to the wp_list_pages
exclusions list via wp_list_pages_excludes