Bug #2215 - Gradebook: cannot edit fname lname for student (and probably editing name doesn't make sense anyway)
Fully removing the ability to edit a students first and last name - the
legacy Gradebook functionality allowed for modification of a student users
first and last name meta. Because this meta is filtered for OpenLab
Xprofile fields, it makes this kind of updating confusing. Furthermore,
instructors should not have this capability anyway.

This feature may come back in the future as OpenLab Gradebook-only
custom first/last fields.