Bug #2241 - Gradebook: when plugin is activated, prompt to create a gradebook
When a user first activates the OpenLab Gradebook plugin, and goes to the
OpenLab Gradebook menu item, they will see a modal asking if they want to
create a new Gradebook. This ticket requested that the name placeholder
for that initial modal be the related course name - in order to provide
that functionaltiy in an agnostic way, the placeholder is set up via a
filter. In Mu Plugins, this filter is hooked into and provided with the
related course name (if available).