Feature #2260 - Gradebook: transition to using pre-exisiting function for finding group id by blog id
The functions involved in this transition are in mu-plugins, however
openlab_get_group_id_by_blog_id() is inside a plugin (WDS Citytech). It's highly unlikely WDS
Citytech
would be deactivated, but just to be on the safe side, I created a special mu-plugins function
wrapper that checks for openlab_get_group_id_by_blog_id(), and uses a fallback if that function
is not available.