Bug #2638 - WP Nav adjustment - truncate long site names
This is for the site name in the WP Admin Bar on group sites. The client-side
trunction was not properly firing on those site names - however the text-overflow CSS
property is now well supported, and so instead of attempting to resolve the issue
within the client-side script, I simply updated the site name to leverage the
text-overflow property.