Change strategy for expanding room above block editor to account for OL toolbar.
Instead of changing the flow using relative positioning as introduced
in 9644d4a, we instead specify the increased size of the toolbar in the CSS.

This change allows the continued use of the 'Top Toolbar' feature in the
Block Editor, which fixes the toolbar at the top of the viewport.

See #2689, #2668.