Ensure that blog_public setting is properly filled during group creation.
Since the move to asynchronous steps for group creation, it's not always
guaranteed that the group site will be created by the time the second step
of group creation is loaded. As such, we need to have an intelligent fallback
when rendering the default blog_public option on that second step. See #2610.

See #2818.