More refinement to the display of group Acknowledgements.
When a group's Creators are different from its contacts, and when that
group also has relevant clone history to display, both pieces of information
should be displayed in the Acknowledgements section. This requires the
markup to be generated in a way that accounts for the possibility of multiple
credits "chunks".

See #2934.