Don't pull only 'autoload' options when cloning sites.
Some important options have 'autoload=no'. In particular, some versions of
gravityforms seem to have set gf_db_version set to autoload=no, which caused
it not to be copied during cloning, which in turn caused forms not to show
up (because GF thought it was looking at different database schema)

See #3012.