Fix the auto-addition of Group Profile and Home links to new sites.
After creating a site by coping from a template, we add nav menu items
for 'Group Profile' and 'Home' to the primary nav menu of the new site.
This functionality was broken by a change in WordPress: When creating
nav items via wp_update_nav_menu_object(), a call to `term_exists()` is
made when associating the new nav_menu_item object with the corresponding
nav_menu taxonomy term. Previously, this `term_exists()` call used its
own bespoke SQL statement to make this query, and as such, it was always
successful at finding the newly-created nav_menu term. After a recent
WordPress update, `term_exists()` uses `WP_Term_Query` internally, which
means that it also receives any filters related to `WP_Term_Query`. In
our case, the taxonomy-terms-order plugin that runs on the main site
(but not subsites) was adding an `ORDER BY t.term_order` clause to the
`term_exists()` query, a clause that depends on the modification to the
`wp_terms` database table schema performed by taxonomy-terms-order. Since
this modification hadn't taken place on the terms table on the new site,
and the `ORDER BY t.term_order` clause always caused an error. As such,
the `term_exists()` query failed, and the newly created `nav_menu_item`
term was not associated with the nav menu.

We work around this problem by unhooking the taxonomy-terms-order filter
when creating the nav menu items during site creation.

See #3089.