Ignore bookly 'backend' directory when determining PHP compat.
The plugin calls static class methods from inside template files, which
are included from inside class methods. PHPCS says this is a late
static binding, but to me it just looks like a scope issue. For the time
being we will skip this, because it doesn't seem to be a PHP version issue.