Hotfix for highlighter-pro PHP 8 incompatibility.
The plugin's Redux_Filesystem class is a singleton, invoked using
a `get_instance()` method. An optional `$parent` class can be passed
to `get_instance()`, which is added to the `Redux_Filesystem` class
after it's invoked. But the `__construct()` method tries to operate
on this `$parent` property, which takes place before the `$parent`
property has been set in `get_instance()`.

See #3129.