First pass at a port of the CBOX-OL site-template feature.
In an effort to make this feature easy to update from the upstream
CBOX-OL in the future, the primary library files from CBOX-OL are copied
over, mostly unmodified. The one exception is a few constants and paths
that are modified in the site-templates.php file. Built JS and CSS files
are totally unmodified.

In order to make this work, we need to do some magic:

- We polyfill some CBOX-OL core functions and classes, including a mock of the GroupType class.
- We reproduce some site-template-related functionality from CBOX-OL that doesn't live in the site-templates.php file.
- We provide a standalone 'Group Type Templates' dashboard panel (under Site Templates) where admins can manage the default template for each group type.
- We do a small amount of manipulation using hidden fields in order to fully mimic the mockup and the behavior of the group create/admin panel from CBOX-OL, in particular where the #set-up-site-toggle ID in CBOX-OL differs from #wds_website_check in OpenLab.

See #3161.