Remove redundant 'non-City-Tech' signup code check during registration.
Placing the logic here means that the field is checked once the 'signup_validation_code'
element is added to the DOM, which can happen if you first enter a non-CT email
and then replace it with a CT one. The logic is redundant anyway, since validation
takes place later in the same function.

See #3245.