Mods to the way that contact form markup is wrapped.
Previously, the contact form in the 'help' section was built
using hardcoded form references. This was removed in 18c40451297b168e03dde6fb489cfc4cd6e5743d
as part of work on #3192. However, this had the side-effect of
breaking other instances of the contact form. This commit works
around this issue by providing special shortcode preprocessing
for the form at about/contact-us.

See #3289.