New approach for excluding off-limits docs from queries.
This new approach is necessary because of the limitations of
WP_Query that don't allow you to use 'OR' relations between
the meta_query and the author__not_in parameters. Instead,
we have to use more explicit logic to pull up a list of off-limits
docs, then loop through them to remove items belonging to the user.

See #3174.